Videos CRUD operations

+++ server/abl/video/delete-abl.js
@@ -1,12 +1,23 @@
 const path = require("path");
 const Ajv = require("ajv").default;
+const addFormats = require("ajv-formats")
+const ajv = new Ajv();
+addFormats(ajv);
 const VideoDao = require("../../dao/video-dao")
+const schema = require("../../validation/reqParamSchema");
 let dao = new VideoDao(
     path.join(__dirname,"..","..","storage","videos.json")
 );
 
 async function DeleteAbl(req, res){
-    // TODO validate that id param is set?
+    if (!ajv.validate(schema, req.params)){
+        res.status(400).json({
+            message: "Invalid params",
+            reason: ajv.errors
+        });
+
+        return;
+    }
     let deleted = await dao.delete(req.params.id);
 
     deleted ?
